Api update

+++ data/src/androidMain/kotlin/fobo66/valiutchik/core/model/datasource/FormattingDataSourceImpl.kt
@@ -49,7 +49,7 @@ class FormattingDataSourceImpl(
         Currency.getInstance(BYN)
     }
 
-    override fun formatCurrencyValue(value: Double): String = NumberFormatter
+    override fun formatCurrencyValue(value: Float): String = NumberFormatter
         .withLocale(locale)
         .unit(targetCurrency)
         .format(value)
